requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
notion-client==3.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==0.52.1
sxtwl==2.0.7
typing_extensions==4.15.0
typing_inspection==0.4.2
uvicorn==0.41.0
